requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
Pillow>=10.0.0
plotly>=5.15.0
streamlit-option-menu>=0.3.6 